safari-534.53-branch/LayoutTests/http/tests/security/view-source-no-javascript-url.html

<html>
<body>
<script>
if (window.layoutTestController)
  layoutTestController.dumpAsText();
</script>
<script>
function loaded() {
  frames[0].location = "javascript:alert('FAIL');";
}
</script>
<p>This test passes if it does not alert FAIL.</p>
<iframe viewsource src="resources/innocent-victim.html" onload="loaded()">
</iframe>
</body>
</html>
