LayoutTests/http/tests/security/view-source-no-javascript-url-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/resources/innocent-victim.html from frame with URL http://127.0.0.1:8000/security/view-source-no-javascript-url.html. Domains, protocols and ports must match.

This test passes if it does not alert FAIL.
